[ESWE] Adds logic to create and retrieve a job (#38)
This PR makes several significant changes. It adds a new `Job` endpoint
and also does some refactoring.

It also starts a folder structure refactor that is more aligned with the
principles of Clean Architecture and DDD. After the CMS logic is
integrated with the front end, the entire refactor will be finished.