[ESWE-960] Get job contains createdAt attribute (#44)
This PR adds the audit attribute `createdAt` when a Job is retrieved.

It also adds more technical debt that must be addressed in the following
deliveries. That debt has increased in how the code manages the fake
data as requests, responses, objects and time.

A proper refactor and use of proper creational patterns, like factories
or Object Mother, will reduce duplication and make our tests simpler.