[ESWE-962] Filter jobs by title or employer name and Improves Job list items schema (#47)
This PR:
- Adds the capability to search, using a single string parameter, any
record whose job title or employer name matches or contains that string.
It doesn't care about text cases, so it will work with capital, regular,
or a mix of the two.
- Changes the schema of the Job list. Now, it returns the minimum
attributes necessary to show, plus the missing Employer name.
- Refactor Job and Employer primary Integration test names so they are
now more consistent and expressive.

Again, it adds a bit more duplication (aka tech debt), which makes sense
as we're discovering the domain. A proper refactor of the elements used
for the test will simplify the code.