[ESWE-962] Fixes the jobs list filter logic (#53)
The expected logic condition of the Jobs list search must be:

`Job sector AND (JobTitle OR Employer name)`

However, the result was not expected because of the limitations of the
JPA Query Methods. We observed instead the following behaviour:

`Job sector AND JobTitle OR Employer name`

The existing tests didn't catch the unexpected behaviour and also needed
a revisit.

This PR fixes the logic and the tests. It replaces the existing JPA
Query Method with a more specific Spring Data JPA @Query that allows the
desired logic condition.