[ESWE-865] Extracts Employer test objects (#80)
This PR comprises a small series of refactors on the testing Employers
domain. The idea is to have the dummy Employer objects that we use and
their representations, like the request body needed for creation or the
response body that we expect to receive from the endpoint, in the same
place.

Another goal is to start abstracting the language used in our
Integration tests so they do not reveal unnecessary implementation
details.