[ESWE-865] Updating null postcode coordinates (#106)
If the OS Places API service is down, the coordinates for a postcode
won't be available. In that case, we'll store null values.

This PR provides a self-healing feature. Every time a postcode is
retrieved when saving or updating a Job, if the code has its coordinates
to null, the service will try to recover them again.