[ESWE-865] Matching candidate job list search radius (#125)
This PR allows for applying a Search Radius filter to the Candidate
Matching endpoint jobs list.

NOTE: After rebasing the main branch, I found that changes and
refactoring have been made to the testing code, affecting how distance
and postcodes are used. Thus, two different ways of setting the
scenarios now live together. At some point, a deep refactoring will be
needed to set up the scenarios and the structure of the tests.