[ESWE-1137] 'No fixed abode' release area postcode and any Search Radius should return Job list (#133)
This PR fixes an unwanted behaviour on the backend Matching Candidates
endpoint for prisoners with a special 'No fixed abode' release area
postcode (NF1 1NF) assigned and any Search Radius provided. With these
two values passed, the result will always be an empty list. It's
interesting to note that if Search Radius is not provided, the list is
generated correctly.

In summary, when a postcode has no associated coordinates for any
reason, the Search Radius filter must be avoided and should not affect
the filtering process.

This PR solves this behaviour by also checking that the coordinates
related to the release area are not null.